Create script.py
A python script that outputs the schools connected and schools not connected to roads in Dagoretti South constituency.

DATA:
constituency shapefile
roads shapefile
schools shapefile

Geoprocessing tool used is BUFFER where schools intersecting with a 15m road are considered connected to a road whereas those not intersecting with the 15m road buffer are considered not connected.